// libs/modules/tasks/src/lib/exceptions/error-codes.enum.ts
export enum TaskErrorCode {
  TASK_NOT_FOUND = 'TASK_NOT_FOUND',
  TASK_LABEL_NOT_FOUND = 'TASK_LABEL_NOT_FOUND',
  TASK_LABEL_ALREADY_EXISTS = 'TASK_LABEL_ALREADY_EXISTS',
  TASK_COMMENT_NOT_FOUND = 'TASK_COMMENT_NOT_FOUND',
  TASK_DEPENDENCY_NOT_FOUND = 'TASK_DEPENDENCY_NOT_FOUND',
  TASK_DEPENDENCY_CYCLE_DETECTED = 'TASK_DEPENDENCY_CYCLE_DETECTED',
  TASK_DEPENDENCY_ALREADY_EXISTS = 'TASK_DEPENDENCY_ALREADY_EXISTS',
  TASK_SELF_DEPENDENCY_NOT_ALLOWED = 'TASK_SELF_DEPENDENCY_NOT_ALLOWED',
  TASK_BATCH_OPERATION_FAILED = 'TASK_BATCH_OPERATION_FAILED',
  TASK_INVALID_PARENT = 'TASK_INVALID_PARENT',
  TASK_UNAUTHORIZED = 'TASK_UNAUTHORIZED',
}